backend/requirements.txt

fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.3.0
azure-cognitiveservices-speech==1.31.0
azure-ai-textanalytics==5.3.0
requests==2.31.0
